linux: NVIDIA GDM wayland suspend fix
This is the final solution of what I was able to find and experiment on.
It should solve all the suspend issues under GNOME Wayland sessions with
NVIDIA proprietary drivers.

If this fails, I'll throw this abomination of a GPU out of my fucking
window. Nvidia never fails to disappoint.